Consolidated EffectChain and EffectChainSlot
- EffectChain and EffectChainSlot had no clear separation of responsibility
- Had to keep the duplicated state in both the classes in sync and thus making it more susceptible to bugs

  NOTE: XML saving/loading of effects is broken and will be re-implemented in the new architecture
