+++ src/effects/effectchainslot.cpp
@@ -1,22 +1,40 @@
 #include "effects/effectchainslot.h"
 
-#include "effects/effectrack.h"
-#include "effects/effectxmlelements.h"
-#include "effects/effectslot.h"
+#include "control/controlencoder.h"
 #include "control/controlpotmeter.h"
 #include "control/controlpushbutton.h"
-#include "control/controlencoder.h"
+#include "effects/effectprocessor.h"
+#include "effects/effectrack.h"
+#include "effects/effectslot.h"
+#include "effects/effectsmanager.h"
+#include "effects/effectxmlelements.h"
+#include "engine/effects/engineeffectchain.h"
+#include "engine/effects/engineeffectrack.h"
+#include "engine/effects/message.h"
+#include "engine/engine.h"
 #include "mixer/playermanager.h"
+#include "util/defs.h"
 #include "util/math.h"
+#include "util/sample.h"
 #include "util/xml.h"
 
+
 EffectChainSlot::EffectChainSlot(EffectRack* pRack, const QString& group,
-                                 unsigned int iChainNumber)
+                                 unsigned int iChainNumber,
+                                 EffectsManager* pEffectsManager,
+                                 const QString& id)
         : m_iChainSlotNumber(iChainNumber),
           // The control group names are 1-indexed while internally everything
           // is 0-indexed.
           m_group(group),
-          m_pEffectRack(pRack) {
+          m_pEffectRack(pRack),
+          m_pEffectsManager(pEffectsManager),
+          m_id(id),
+          m_name(""),
+          m_description(""),
+          m_pEngineEffectChain(nullptr) {
+    // qDebug() << "EffectChainSlot::EffectChainSlot " << pRack << ' ' << group << ' ' << iChainNumber;
+
     m_pControlClear = new ControlPushButton(ConfigKey(m_group, "clear"));
     connect(m_pControlClear, SIGNAL(valueChanged(double)),
             this, SLOT(slotControlClear(double)));
@@ -29,6 +47,9 @@ EffectChainSlot::EffectChainSlot(EffectRack* pRack, const QString& group,
 
     m_pControlChainLoaded = new ControlObject(ConfigKey(m_group, "loaded"));
     m_pControlChainLoaded->setReadOnly();
+    if (id != QString()) {
+        m_pControlChainLoaded->forceSet(1.0);
+    }
 
     m_pControlChainEnabled = new ControlPushButton(ConfigKey(m_group, "enabled"));
     m_pControlChainEnabled->setButtonMode(ControlPushButton::POWERWINDOW);
@@ -36,12 +57,12 @@ EffectChainSlot::EffectChainSlot(EffectRack* pRack, const QString& group,
     m_pControlChainEnabled->setDefaultValue(true);
     m_pControlChainEnabled->set(true);
     connect(m_pControlChainEnabled, SIGNAL(valueChanged(double)),
-            this, SLOT(slotControlChainEnabled(double)));
+            this, SLOT(sendParameterUpdate()));
 
     m_pControlChainMix = new ControlPotmeter(ConfigKey(m_group, "mix"), 0.0, 1.0,
                                              false, true, false, true, 1.0);
     connect(m_pControlChainMix, SIGNAL(valueChanged(double)),
-            this, SLOT(slotControlChainMix(double)));
+            this, SLOT(sendParameterUpdate()));
 
     m_pControlChainSuperParameter = new ControlPotmeter(ConfigKey(m_group, "super1"), 0.0, 1.0);
     connect(m_pControlChainSuperParameter, SIGNAL(valueChanged(double)),
@@ -53,7 +74,7 @@ EffectChainSlot::EffectChainSlot(EffectRack* pRack, const QString& group,
     m_pControlChainMixMode->setButtonMode(ControlPushButton::TOGGLE);
     m_pControlChainMixMode->setStates(static_cast<int>(EffectChainMixMode::NumMixModes));
     connect(m_pControlChainMixMode, SIGNAL(valueChanged(double)),
-            this, SLOT(slotControlChainMixMode(double)));
+            this, SLOT(sendParameterUpdate()));
 
     m_pControlChainNextPreset = new ControlPushButton(ConfigKey(m_group, "next_chain"));
     connect(m_pControlChainNextPreset, SIGNAL(valueChanged(double)),
@@ -86,6 +107,8 @@ EffectChainSlot::EffectChainSlot(EffectRack* pRack, const QString& group,
                                        ConfigKey(m_group, "focused_effect"),
                                        true);
     m_pControlChainFocusedEffect->setButtonMode(ControlPushButton::TOGGLE);
+
+    addToEngine(m_pEffectRack->getEngineEffectRack(), m_iChainSlotNumber);
 }
 
 EffectChainSlot::~EffectChainSlot() {
@@ -113,13 +136,141 @@ EffectChainSlot::~EffectChainSlot() {
     }
 
     m_slots.clear();
-    m_pEffectChain.clear();
+    removeFromEngine(m_pEffectRack->getEngineEffectRack(), m_iChainSlotNumber);
+}
+
+void EffectChainSlot::addToEngine(EngineEffectRack* pRack, int iIndex) {
+    m_pEngineEffectChain = new EngineEffectChain(m_id,
+        m_pEffectsManager->registeredInputChannels(),
+        m_pEffectsManager->registeredOutputChannels());
+    EffectsRequest* pRequest = new EffectsRequest();
+    pRequest->type = EffectsRequest::ADD_CHAIN_TO_RACK;
+    pRequest->pTargetRack = pRack;
+    pRequest->AddChainToRack.pChain = m_pEngineEffectChain;
+    pRequest->AddChainToRack.iIndex = iIndex;
+    m_pEffectsManager->writeRequest(pRequest);
+
+    // Add all effects.
+    for (int i = 0; i < m_effects.size(); ++i) {
+        // Add the effect to the engine.
+        EffectPointer pEffect = m_effects[i];
+        if (pEffect) {
+            pEffect->addToEngine(m_pEngineEffectChain, i, m_enabledInputChannels);
+        }
+    }
+    sendParameterUpdate();
+}
+
+void EffectChainSlot::removeFromEngine(EngineEffectRack* pRack, int iIndex) {
+    // Order doesn't matter when removing.
+    for (int i = 0; i < m_effects.size(); ++i) {
+        EffectPointer pEffect = m_effects[i];
+        if (pEffect) {
+            pEffect->removeFromEngine(m_pEngineEffectChain, i);
+        }
+    }
+
+    EffectsRequest* pRequest = new EffectsRequest();
+    pRequest->type = EffectsRequest::REMOVE_CHAIN_FROM_RACK;
+    pRequest->pTargetRack = pRack;
+    pRequest->RemoveChainFromRack.pChain = m_pEngineEffectChain;
+    pRequest->RemoveChainFromRack.iIndex = iIndex;
+    m_pEffectsManager->writeRequest(pRequest);
+
+    m_pEngineEffectChain = nullptr;
+}
+
+const QString& EffectChainSlot::name() const {
+    return m_name;
+}
+
+void EffectChainSlot::setName(const QString& name) {
+    m_name = name;
+    emit(updated());
+}
+
+QString EffectChainSlot::description() const {
+    return m_description;
+}
+
+void EffectChainSlot::setDescription(const QString& description) {
+    m_description = description;
+    emit(updated());
+}
+
+void EffectChainSlot::setMix(const double& dMix) {
+    m_pControlChainMix->set(dMix);
+    sendParameterUpdate();
+}
+
+void EffectChainSlot::addEffect(EffectPointer pEffect) {
+    // qDebug() << debugString() << "addEffect" << pEffect;
+    if (!pEffect) {
+        // Insert empty effects to preserve chain order
+        // when loading chains with empty effects
+        m_effects.append(pEffect);
+        return;
+    }
+
+    VERIFY_OR_DEBUG_ASSERT(!m_effects.contains(pEffect)) {
+        return;
+    }
+
+    m_effects.append(pEffect);
+    pEffect->addToEngine(m_pEngineEffectChain, m_effects.size() - 1, m_enabledInputChannels);
+    slotChainEffectChanged(m_effects.size() - 1);
+}
+
+void EffectChainSlot::replaceEffect(unsigned int effectSlotNumber,
+                                EffectPointer pEffect) {
+    // qDebug() << debugString() << "replaceEffect" << effectSlotNumber << pEffect;
+    while (effectSlotNumber >= static_cast<unsigned int>(m_effects.size())) {
+        if (pEffect.isNull()) {
+            return;
+        }
+        m_effects.append(EffectPointer());
+    }
+
+    EffectPointer pOldEffect = m_effects[effectSlotNumber];
+    if (!pOldEffect.isNull()) {
+        pOldEffect->removeFromEngine(m_pEngineEffectChain, effectSlotNumber);
+    }
+
+    m_effects.replace(effectSlotNumber, pEffect);
+    if (!pEffect.isNull()) {
+        pEffect->addToEngine(m_pEngineEffectChain, effectSlotNumber, m_enabledInputChannels);
+    }
+
+    slotChainEffectChanged(effectSlotNumber);
+}
+
+void EffectChainSlot::removeEffect(unsigned int effectSlotNumber) {
+    replaceEffect(effectSlotNumber, EffectPointer());
+}
+
+void EffectChainSlot::refreshAllEffects() {
+    for (int i = 0; i < m_effects.size(); ++i) {
+        slotChainEffectChanged(i);
+    }
+}
+
+const QList<EffectPointer>& EffectChainSlot::effects() const {
+    return m_effects;
+}
+
+void EffectChainSlot::sendParameterUpdate() {
+    EffectsRequest* pRequest = new EffectsRequest();
+    pRequest->type = EffectsRequest::SET_EFFECT_CHAIN_PARAMETERS;
+    pRequest->pTargetChain = m_pEngineEffectChain;
+    pRequest->SetEffectChainParameters.enabled = m_pControlChainEnabled->get();
+    pRequest->SetEffectChainParameters.mix_mode = static_cast<EffectChainMixMode>(
+                                                    static_cast<int>(m_pControlChainMixMode->get()));
+    pRequest->SetEffectChainParameters.mix = m_pControlChainMix->get();
+    m_pEffectsManager->writeRequest(pRequest);
 }
 
 QString EffectChainSlot::id() const {
-    if (m_pEffectChain)
-        return m_pEffectChain->id();
-    return "";
+    return m_id;
 }
 
 double EffectChainSlot::getSuperParameter() const {
@@ -135,168 +286,41 @@ void EffectChainSlot::setSuperParameterDefaultValue(double value) {
     m_pControlChainSuperParameter->setDefaultValue(value);
 }
 
-void EffectChainSlot::slotChainNameChanged(const QString&) {
-    emit(updated());
-}
+void EffectChainSlot::slotChainEffectChanged(unsigned int effectSlotNumber) {
+    qDebug() << debugString() << "slotChainEffectChanged" << effectSlotNumber;
+    EffectSlotPointer pSlot;
+    EffectPointer pEffect;
 
-void EffectChainSlot::slotChainEnabledChanged(bool bEnabled) {
-    m_pControlChainEnabled->set(bEnabled);
-    emit(updated());
-}
-
-void EffectChainSlot::slotChainMixChanged(double mix) {
-    m_pControlChainMix->set(mix);
-    emit(updated());
-}
-
-void EffectChainSlot::slotChainMixModeChanged(EffectChainMixMode mixMode) {
-    m_pControlChainMixMode->set(static_cast<double>(mixMode));
-    emit(updated());
-}
-
-void EffectChainSlot::slotChainChannelStatusChanged(const QString& group,
-                                                    bool enabled) {
-    ChannelInfo* pInfo = m_channelInfoByName.value(group, NULL);
-    if (pInfo != NULL && pInfo->pEnabled != NULL) {
-        pInfo->pEnabled->set(enabled);
-        emit(updated());
+    if (m_effects.size() > m_slots.size()) {
+        qWarning() << debugString() << "has too few slots for effect";
     }
-}
-
-void EffectChainSlot::slotChainEffectChanged(unsigned int effectSlotNumber,
-                                             bool shouldEmit) {
-    //qDebug() << debugString() << "slotChainEffectChanged" << effectSlotNumber;
-    if (m_pEffectChain) {
-        const QList<EffectPointer> effects = m_pEffectChain->effects();
-        EffectSlotPointer pSlot;
-        EffectPointer pEffect;
-
-        if (effects.size() > m_slots.size()) {
-            qWarning() << debugString() << "has too few slots for effect";
-        }
-
-        if (effectSlotNumber < (unsigned) m_slots.size()) {
-            pSlot = m_slots.at(effectSlotNumber);
-        }
-        if (effectSlotNumber < (unsigned) effects.size()) {
-            pEffect = effects.at(effectSlotNumber);
-        }
-        if (pSlot != nullptr) {
-            pSlot->loadEffect(pEffect, m_pEffectRack->isAdoptMetaknobValueEnabled());
-        }
-
-        m_pControlNumEffects->forceSet(math_min(
-                static_cast<unsigned int>(m_slots.size()),
-                m_pEffectChain->numEffects()));
-
-        if (shouldEmit) {
-            emit(updated());
-        }
+    if (effectSlotNumber < (unsigned) m_slots.size()) {
+        pSlot = m_slots.at(effectSlotNumber);
     }
-}
-
-void EffectChainSlot::loadEffectChainToSlot(EffectChainPointer pEffectChain) {
-    //qDebug() << debugString() << "loadEffectChainToSlot" << (pEffectChain ? pEffectChain->id() : "(null)");
-    clear();
-
-    if (pEffectChain) {
-        m_pEffectChain = pEffectChain;
-
-        connect(m_pEffectChain.data(), SIGNAL(effectChanged(unsigned int)),
-                this, SLOT(slotChainEffectChanged(unsigned int)));
-        connect(m_pEffectChain.data(), SIGNAL(nameChanged(const QString&)),
-                this, SLOT(slotChainNameChanged(const QString&)));
-        connect(m_pEffectChain.data(), SIGNAL(enabledChanged(bool)),
-                this, SLOT(slotChainEnabledChanged(bool)));
-        connect(m_pEffectChain.data(), SIGNAL(mixChanged(double)),
-                this, SLOT(slotChainMixChanged(double)));
-        connect(m_pEffectChain.data(), SIGNAL(mixModeChanged(EffectChainMixMode)),
-                this, SLOT(slotChainMixModeChanged(EffectChainMixMode)));
-        connect(m_pEffectChain.data(), SIGNAL(channelStatusChanged(const QString&, bool)),
-                this, SLOT(slotChainChannelStatusChanged(const QString&, bool)));
-
-        m_pControlChainLoaded->forceSet(true);
-        m_pControlChainMixMode->set(
-                static_cast<double>(m_pEffectChain->mixMode()));
-
-        // Mix and enabled channels are persistent properties of the chain slot,
-        // not of the chain. Propagate the current settings to the chain.
-        m_pEffectChain->setMix(m_pControlChainMix->get());
-        m_pEffectChain->setEnabled(m_pControlChainEnabled->get() > 0.0);
-
-        // Don't emit because we will below.
-        for (int i = 0; i < m_slots.size(); ++i) {
-            slotChainEffectChanged(i, false);
-        }
+    if (effectSlotNumber < (unsigned) m_effects.size()) {
+        pEffect = m_effects.at(effectSlotNumber);
+    }
+    if (pSlot != nullptr) {
+        pSlot->loadEffect(pEffect, m_pEffectRack->isAdoptMetaknobValueEnabled());
     }
 
-    emit(effectChainLoaded(pEffectChain));
-    emit(updated());
-}
-
-void EffectChainSlot::updateRoutingSwitches() {
-    VERIFY_OR_DEBUG_ASSERT(m_pEffectChain) {
-        return;
-    }
-    for (const ChannelInfo* pChannelInfo : m_channelInfoByName) {
-        if (pChannelInfo->pEnabled->toBool()) {
-            m_pEffectChain->enableForInputChannel(pChannelInfo->handle_group);
-        } else {
-            m_pEffectChain->disableForInputChannel(pChannelInfo->handle_group);
-        }
-    }
-}
-
-EffectChainPointer EffectChainSlot::getEffectChain() const {
-    return m_pEffectChain;
-}
-
-EffectChainPointer EffectChainSlot::getOrCreateEffectChain(
-        EffectsManager* pEffectsManager) {
-    if (!m_pEffectChain) {
-        EffectChainPointer pEffectChain(
-                new EffectChain(pEffectsManager, QString()));
-        //: Name for an empty effect chain, that is created after eject
-        pEffectChain->setName(tr("Empty Chain"));
-        loadEffectChainToSlot(pEffectChain);
-        pEffectChain->addToEngine(m_pEffectRack->getEngineEffectRack(), m_iChainSlotNumber);
-        pEffectChain->updateEngineState();
-        updateRoutingSwitches();
-    }
-    return m_pEffectChain;
+    m_pControlNumEffects->forceSet(math_min(
+            static_cast<unsigned int>(m_slots.size()),
+            static_cast<unsigned int>(m_effects.size())));
 }
 
 void EffectChainSlot::clear() {
-    // Stop listening to signals from any loaded effect
-    if (m_pEffectChain) {
-        m_pEffectChain->removeFromEngine(m_pEffectRack->getEngineEffectRack(),
-                                         m_iChainSlotNumber);
-        for (EffectSlotPointer pSlot : m_slots) {
-            pSlot->clear();
-        }
-        m_pEffectChain->disconnect(this);
-        m_pEffectChain.clear();
-    }
     m_pControlNumEffects->forceSet(0.0);
     m_pControlChainLoaded->forceSet(0.0);
     m_pControlChainMixMode->set(
             static_cast<double>(EffectChainMixMode::DrySlashWet));
-    emit(updated());
-}
-
-unsigned int EffectChainSlot::numSlots() const {
-    //qDebug() << debugString() << "numSlots";
-    return m_slots.size();
 }
 
 EffectSlotPointer EffectChainSlot::addEffectSlot(const QString& group) {
-    //qDebug() << debugString() << "addEffectSlot" << group;
+    // qDebug() << debugString() << "addEffectSlot" << group;
 
     EffectSlot* pEffectSlot = new EffectSlot(group, m_iChainSlotNumber,
                                              m_slots.size());
-    // Rebroadcast effectLoaded signals
-    connect(pEffectSlot, SIGNAL(effectLoaded(EffectPointer, unsigned int)),
-            this, SLOT(slotEffectLoaded(EffectPointer, unsigned int)));
     connect(pEffectSlot, SIGNAL(clearEffect(unsigned int)),
             this, SLOT(slotClearEffect(unsigned int)));
     connect(pEffectSlot, SIGNAL(nextEffect(unsigned int, unsigned int, EffectPointer)),
@@ -327,23 +351,22 @@ void EffectChainSlot::registerInputChannel(const ChannelHandleAndGroup& handle_g
             ConfigKey(m_group, QString("group_%1_enable").arg(handle_group.name())),
             true, initialValue);
     pEnableControl->setButtonMode(ControlPushButton::POWERWINDOW);
+    if (pEnableControl->toBool()) {
+        enableForInputChannel(handle_group);
+    }
 
     ChannelInfo* pInfo = new ChannelInfo(handle_group, pEnableControl);
     m_channelInfoByName[handle_group.name()] = pInfo;
+
+    // m_channelStatusMapper will emit a mapped(handle_group.name()) signal whenever
+    // the valueChanged(double) signal is emitted by pEnableControl 
     m_channelStatusMapper.setMapping(pEnableControl, handle_group.name());
     connect(pEnableControl, SIGNAL(valueChanged(double)),
             &m_channelStatusMapper, SLOT(map()));
 }
 
-void EffectChainSlot::slotEffectLoaded(EffectPointer pEffect, unsigned int slotNumber) {
-    // const int is a safe read... don't bother locking
-    emit(effectLoaded(pEffect, m_iChainSlotNumber, slotNumber));
-}
-
 void EffectChainSlot::slotClearEffect(unsigned int iEffectSlotNumber) {
-    if (m_pEffectChain) {
-        m_pEffectChain->removeEffect(iEffectSlotNumber);
-    }
+    removeEffect(iEffectSlotNumber);
 }
 
 EffectSlotPointer EffectChainSlot::getEffectSlot(unsigned int slotNumber) {
@@ -357,65 +380,33 @@ EffectSlotPointer EffectChainSlot::getEffectSlot(unsigned int slotNumber) {
 
 void EffectChainSlot::slotControlClear(double v) {
     if (v > 0) {
+        for (EffectSlotPointer pSlot : m_slots) {
+            pSlot->clear();
+        }
         clear();
     }
 }
 
-void EffectChainSlot::slotControlChainEnabled(double v) {
-    //qDebug() << debugString() << "slotControlChainEnabled" << v;
-    if (m_pEffectChain) {
-        m_pEffectChain->setEnabled(v > 0);
-    }
-}
-
-void EffectChainSlot::slotControlChainMix(double v) {
-    //qDebug() << debugString() << "slotControlChainMix" << v;
-
-    // Clamp to [0.0, 1.0]
-    if (v < 0.0 || v > 1.0) {
-        qWarning() << debugString() << "value out of limits";
-        v = math_clamp(v, 0.0, 1.0);
-        m_pControlChainMix->set(v);
-    }
-    if (m_pEffectChain) {
-        m_pEffectChain->setMix(v);
-    }
-}
-
 void EffectChainSlot::slotControlChainSuperParameter(double v, bool force) {
-    //qDebug() << debugString() << "slotControlChainSuperParameter" << v;
+    // qDebug() << debugString() << "slotControlChainSuperParameter" << v;
 
-    // Clamp to [0.0, 1.0]
-    if (v < 0.0 || v > 1.0) {
-        qWarning() << debugString() << "value out of limits";
-        v = math_clamp(v, 0.0, 1.0);
-        m_pControlChainSuperParameter->set(v);
-    }
+    m_pControlChainSuperParameter->set(v);
     for (const auto& pSlot : m_slots) {
         pSlot->setMetaParameter(v, force);
     }
 }
 
-void EffectChainSlot::slotControlChainMixMode(double v) {
-    // Intermediate cast to integer is needed for VC++.
-    EffectChainMixMode type = static_cast<EffectChainMixMode>(int(v));
-    (void)v; // this avoids a false warning with g++ 4.8.1
-    if (m_pEffectChain && type < EffectChainMixMode::NumMixModes) {
-        m_pEffectChain->setMixMode(type);
-    }
-}
-
 void EffectChainSlot::slotControlChainSelector(double v) {
-    //qDebug() << debugString() << "slotControlChainSelector" << v;
-    if (v > 0) {
-        emit(nextChain(m_iChainSlotNumber, m_pEffectChain));
-    } else if (v < 0) {
-        emit(prevChain(m_iChainSlotNumber, m_pEffectChain));
-    }
+//     qDebug() << debugString() << "slotControlChainSelector" << v;
+//     if (v > 0) {
+//         emit(nextChain(m_iChainSlotNumber, m_pEffectChain));
+//     } else if (v < 0) {
+//         emit(prevChain(m_iChainSlotNumber, m_pEffectChain));
+//     }
 }
 
 void EffectChainSlot::slotControlChainNextPreset(double v) {
-    //qDebug() << debugString() << "slotControlChainNextPreset" << v;
+    // qDebug() << debugString() << "slotControlChainNextPreset" << v;
     if (v > 0) {
         slotControlChainSelector(1);
     }
@@ -429,17 +420,75 @@ void EffectChainSlot::slotControlChainPrevPreset(double v) {
 }
 
 void EffectChainSlot::slotChannelStatusChanged(const QString& group) {
-    if (m_pEffectChain) {
-        ChannelInfo* pChannelInfo = m_channelInfoByName.value(group, NULL);
-        if (pChannelInfo != NULL && pChannelInfo->pEnabled != NULL) {
-            bool bEnable = pChannelInfo->pEnabled->toBool();
-            if (bEnable) {
-                m_pEffectChain->enableForInputChannel(pChannelInfo->handle_group);
-            } else {
-                m_pEffectChain->disableForInputChannel(pChannelInfo->handle_group);
-            }
+    ChannelInfo* pChannelInfo = m_channelInfoByName.value(group, NULL);
+    if (pChannelInfo != NULL && pChannelInfo->pEnabled != NULL) {
+        bool bEnable = pChannelInfo->pEnabled->toBool();
+        if (bEnable) {
+            enableForInputChannel(pChannelInfo->handle_group);
+        } else {
+            disableForInputChannel(pChannelInfo->handle_group);
         }
     }
+}
+
+void EffectChainSlot::enableForInputChannel(const ChannelHandleAndGroup& handle_group) {
+    if (m_enabledInputChannels.contains(handle_group)) {
+        return;
+    }
+
+    EffectsRequest* request = new EffectsRequest();
+    request->type = EffectsRequest::ENABLE_EFFECT_CHAIN_FOR_INPUT_CHANNEL;
+    request->pTargetChain = m_pEngineEffectChain;
+    request->EnableInputChannelForChain.pChannelHandle = &handle_group.handle();
+
+    // Allocate EffectStates here in the main thread to avoid allocating
+    // memory in the realtime audio callback thread. Pointers to the
+    // EffectStates are passed to the EffectRequest and the EffectProcessorImpls
+    // store the pointers. The containers of EffectState* pointers get deleted
+    // by ~EffectsRequest, but the EffectStates are managed by EffectProcessorImpl.
+    auto pEffectStatesMapArray = new EffectStatesMapArray;
+
+    //TODO: get actual configuration of engine
+    const mixxx::EngineParameters bufferParameters(
+          mixxx::AudioSignal::SampleRate(96000),
+          MAX_BUFFER_LEN / mixxx::kEngineChannelCount);
+
+    for (int i = 0; i < m_effects.size(); ++i) {
+        auto& statesMap = (*pEffectStatesMapArray)[i];
+        if (m_effects[i] != nullptr) {
+            for (const auto& outputChannel : m_pEffectsManager->registeredOutputChannels()) {
+                if (kEffectDebugOutput) {
+                    qDebug() << debugString() << "EffectChain::enableForInputChannel creating EffectState for input" << handle_group << "output" << outputChannel;
+                }
+                statesMap.insert(outputChannel.handle(),
+                        m_effects[i]->createState(bufferParameters));
+            }
+        } else {
+            for (EffectState* pState : statesMap) {
+                if (pState != nullptr) {
+                    delete pState;
+                }
+            }
+            statesMap.clear();
+        }
+    }
+    request->EnableInputChannelForChain.pEffectStatesMapArray = pEffectStatesMapArray;
+
+    m_pEffectsManager->writeRequest(request);
+
+    m_enabledInputChannels.insert(handle_group);
+}
+
+void EffectChainSlot::disableForInputChannel(const ChannelHandleAndGroup& handle_group) {
+    if (!m_enabledInputChannels.remove(handle_group)) {
+        return;
+    }
+
+    EffectsRequest* request = new EffectsRequest();
+    request->type = EffectsRequest::DISABLE_EFFECT_CHAIN_FOR_INPUT_CHANNEL;
+    request->pTargetChain = m_pEngineEffectChain;
+    request->DisableInputChannelForChain.pChannelHandle = &handle_group.handle();
+    m_pEffectsManager->writeRequest(request);
 }
 
 unsigned int EffectChainSlot::getChainSlotNumber() const {
@@ -448,62 +497,58 @@ unsigned int EffectChainSlot::getChainSlotNumber() const {
 
 QDomElement EffectChainSlot::toXml(QDomDocument* doc) const {
     QDomElement chainElement = doc->createElement(EffectXml::Chain);
-    if (m_pEffectChain == nullptr) {
-        // ejected chains are stored empty <EffectChain/>
-        return chainElement;
-    }
 
-    XmlParse::addElement(*doc, chainElement, EffectXml::ChainId,
-            m_pEffectChain->id());
-    XmlParse::addElement(*doc, chainElement, EffectXml::ChainName,
-            m_pEffectChain->name());
-    XmlParse::addElement(*doc, chainElement, EffectXml::ChainDescription,
-            m_pEffectChain->description());
-    XmlParse::addElement(*doc, chainElement, EffectXml::ChainMixMode,
-            EffectChain::mixModeToString(
-                    static_cast<EffectChainMixMode>(
-                            static_cast<int>(m_pControlChainMixMode->get()))));
-    XmlParse::addElement(*doc, chainElement, EffectXml::ChainSuperParameter,
-            QString::number(m_pControlChainSuperParameter->get()));
+    // XmlParse::addElement(*doc, chainElement, EffectXml::ChainId,
+    //         m_pEffectChain->id());
+    // XmlParse::addElement(*doc, chainElement, EffectXml::ChainName,
+    //         m_pEffectChain->name());
+    // XmlParse::addElement(*doc, chainElement, EffectXml::ChainDescription,
+    //         m_pEffectChain->description());
+    // XmlParse::addElement(*doc, chainElement, EffectXml::ChainMixMode,
+    //         EffectChain::mixModeToString(
+    //                 static_cast<EffectChainMixMode>(
+    //                         static_cast<int>(m_pControlChainMixMode->get()))));
+    // XmlParse::addElement(*doc, chainElement, EffectXml::ChainSuperParameter,
+    //         QString::number(m_pControlChainSuperParameter->get()));
 
-    QDomElement effectsElement = doc->createElement(EffectXml::EffectsRoot);
-    for (const auto& pEffectSlot : m_slots) {
-        QDomElement effectNode;
-        if (pEffectSlot->getEffect()) {
-            effectNode = pEffectSlot->toXml(doc);
-        } else {
-            // Create empty element to ensure effects stay in order
-            // if there are empty slots before loaded slots.
-            effectNode = doc->createElement(EffectXml::Effect);
-        }
-        effectsElement.appendChild(effectNode);
-    }
-    chainElement.appendChild(effectsElement);
+    // QDomElement effectsElement = doc->createElement(EffectXml::EffectsRoot);
+    // for (const auto& pEffectSlot : m_slots) {
+    //     QDomElement effectNode;
+    //     if (pEffectSlot->getEffect()) {
+    //         effectNode = pEffectSlot->toXml(doc);
+    //     } else {
+    //         // Create empty element to ensure effects stay in order
+    //         // if there are empty slots before loaded slots.
+    //         effectNode = doc->createElement(EffectXml::Effect);
+    //     }
+    //     effectsElement.appendChild(effectNode);
+    // }
+    // chainElement.appendChild(effectsElement);
 
     return chainElement;
 }
 
 void EffectChainSlot::loadChainSlotFromXml(const QDomElement& effectChainElement) {
-    if (!effectChainElement.hasChildNodes()) {
-        return;
-    }
+    // if (!effectChainElement.hasChildNodes()) {
+    //     return;
+    // }
 
-    // FIXME: mix mode is set in EffectChain::createFromXml
+    // // FIXME: mix mode is set in EffectChain::createFromXml
 
-    m_pControlChainSuperParameter->set(XmlParse::selectNodeDouble(
-                                          effectChainElement,
-                                          EffectXml::ChainSuperParameter));
+    // m_pControlChainSuperParameter->set(XmlParse::selectNodeDouble(
+    //                                       effectChainElement,
+    //                                       EffectXml::ChainSuperParameter));
 
-    QDomElement effectsElement = XmlParse::selectElement(effectChainElement,
-                                                         EffectXml::EffectsRoot);
-    QDomNodeList effectsNodeList = effectsElement.childNodes();
-    for (int i = 0; i < m_slots.size(); ++i) {
-        if (m_slots[i] != nullptr) {
-            QDomNode effectNode = effectsNodeList.at(i);
-            if (effectNode.isElement()) {
-                QDomElement effectElement = effectNode.toElement();
-                m_slots[i]->loadEffectSlotFromXml(effectElement);
-            }
-        }
-    }
+    // QDomElement effectsElement = XmlParse::selectElement(effectChainElement,
+    //                                                      EffectXml::EffectsRoot);
+    // QDomNodeList effectsNodeList = effectsElement.childNodes();
+    // for (int i = 0; i < m_slots.size(); ++i) {
+    //     if (m_slots[i] != nullptr) {
+    //         QDomNode effectNode = effectsNodeList.at(i);
+    //         if (effectNode.isElement()) {
+    //             QDomElement effectElement = effectNode.toElement();
+    //             m_slots[i]->loadEffectSlotFromXml(effectElement);
+    //         }
+    //     }
+    // }
 }
