+++ src/engine/effects/engineeffectchain.cpp
@@ -113,7 +113,8 @@ bool EngineEffectChain::processEffectsRequest(EffectsRequest& message,
             if (kEffectDebugOutput) {
                 qDebug() << debugString() << this << "SET_EFFECT_CHAIN_PARAMETERS"
                          << "enabled" << message.SetEffectChainParameters.enabled
-                         << "mix" << message.SetEffectChainParameters.mix;
+                         << "mix" << message.SetEffectChainParameters.mix
+                         << "mix_mode" << static_cast<int>(message.SetEffectChainParameters.mix_mode);
             }
             response.success = updateParameters(message);
             break;
